hotfix [#2]: Ifs did not inherit their parent block's scope
Fixes a bug where ifs do not inherit their parent's scope. This was caused by varying case within the if execution, class's constructor. In which the scope was, resolved to, or was set to `this.scope`, while the execution referenced either or. Because `this.*` was a member of the accessor of `this`, which would return undefined, it is not either a false boolean or an errored Symbol. This means it would default to true
Resolution: relace globally `this.scope` with `this.Scope` to match case of the tokenizer

Fixes #2