feature [casting]: Added support for casting operator `::`
This update brings support for the casting operator `::` to expressions.
It's definition is as a normal operator but is a infix defined on top
of the `class` defaults. The cast operator checks and ensures that LHS
is a class which is verified by a unique symbol passed to the class
static chain. This is stored in the Class.UID which is then re-verified
as the class is passed back. This avoids any circular dependency issues