tests [made]: made unit tests, currently casting
Created unit tests with mocha and chai, at the moment
only casting actually has a unit test. But as the groundwork
is not established, it should be very quick to create new
unit tests, this should greatly.

Majority of the work was done by @The-Quill, so 90% of this
commit's credit goes to him.