hotfix [#6]: range operator did not work
The range operator did not work for variety of reasons, returning
a MALFORMED_TOKEN error in some cases, and an unabstracted error
in others, this was especially probleatic as the dependencies for
the range function would result in circular dependency errors,
which lie as a bug in the underlying node engine rather than
cheddar itself. This was caused by a lack of the construction of a
CheddarNumber, in which the Array did not expect such a token to
appear as classes are completely abstracted. This means it would
correctly recieve the tokens but not copy them correctly. Perhaps
the range operator will be moved into a `range` class like python
for speed and other benifits. As of now the other error was that
the array would get wrapped twice which was the lack of the spread
operator during the construction i.e. `init(Array, [1,2,3])`
rather than `init(Array, 1, 2, 3)`.

Resolution: request the `Number` dependency JIT. Spread the array

Fixes #6