stdlib [string static]: Created strings in STDLIB
Started on the standard library and created `letters` and `digits`
items on the string's static property. This mainly for creating
testing recursive accessors. Which is part of a push to create a Math
library, which is part of the larger slower push to create the stdlib.
Once the Math library is pushed; along with the IO class which will
ship be either v0.4 or v1, Cheddar should be a production ready-lang.
Additionally proccess/shell built-ins will be needed, but the biggest
delay within Cheddar's development is bugs. A couple people have
volunteered to beta test Cheddar, but only one bug report has been
made :/