refactor: evaluation is now non-destructive
CheddarEval used to be destructive in the sense it
would modify the call stack recieved. This would
be problematic of for loops as they would error
when revaluating in the same context. This has been
fixed by mimicing this behavior with a stack
pointer, this also solves problem with recursion
and allows better optimization with preserving
scope variables