feature [props]: Recursive accessors are a thing
Recursive accessors essentially mean evaluation
recursively accesses accessors to acces the
accessors accessing the accessed proprety which
allows for name spaces, methods and opens a
whole new door for capability by the v1s launch
Functions which will ship in v0.3 or v1 (which
ever comes first). Are really all that is
needed until the standard library is the only
development left to do. Of course this is
ignoring bugs