feature [func]: Created inital function execution
Functions are now executable. They are constructed by passing
their body and a given scope. The scope function then builds a
scope body, referencing the cache to avoid recreating a scope.
Then the function is executed as a JS function. In a later commit
this will support executing a CheddarCodeblock. For now though
only JS functions are supported but that will still go a long way