stdlib [math]: created STDLIB libs for math
Created a couple basic ufnctions for the math library. This
includes various constants, such as `Pi`, `Phi`, `Gamma`, etc.
Functions including primality and fibonacci functions have also
been created (credit to: @ConorOBrien-Foxx). This will be part of
release `v1`. It's not likely that a `v0.3` will be made, but the
standard libarary, will have to undergo rigorious testing for the
coverage to account for the significant changes in this version