stdlib [linked]: Linked STDLIB to REPL
The STDLIB is now accessable through the REPL. Sadly, functions do
not work yet. As in the load but do not have the syntax linking to
actually be functional and execute. This will come in the next
commit, which I will announce version v0.3, the pre-release to the
production-ready v1.0