fix [bool]: Default global bool handling
Booleans now have improved and more reliable global behavior.
Operators lookup on the `Cast` reference for a `Bool` cast and
attempted to use that on the inherited defualt `!` which allows it
to delegate as a reference to a boolean which is used in the bool
target when logical NOTed.

Booleans also now have a standardized NOT eq. behavior which is
done by simply calling the Cast's boolean reference on the class.
The resulting boolean is then inverted and simply passed through
out.