fix [LR #5]: Solves LR in ( E ) + prop
Solved rather critical left recursion bug within property grammar
resulting in property heads as a parenthesized expression being
excluded for quite a while. This was sovles by replacing a manual
find-execute-close method with a grammar-base approach which fakes
returning an expression which the property grammar identifies and
returns to retain compatiblity with the CS and evaluator.

Resolution: move to lexer#grammar

Fixes #5