fix [ternary]: Improper evaluation bug
CheddarExpressionTernaries would improperly execute to ShuntingYard
as the Token inheritance would specify both tokens, alpha, and beta
as part of the _Tokens class on the delegate StatementExpression.
This in turn would result in the beta token being discarded as the
alpha token was referenced but not completely as the beta token was
in fact referenced as the alpha was viewed as the ternary delegate.
ShuntingYard would attempt to reorder the tokens but could not
reference the isExpression property giving the expression that the
expression is not in fact a delegate and is discareded. To address
this, the resolution was to send the delegate as an Expression
token through the wrapper to delegate as a StatementExpression,
resulting in the delegate token to be the Tokens set of alpha, and
beta, and then when passed to ShuntingYard, is correctly parsed