fix [func]: Fixed function recursion
Rather big bug fix but recursion was borked for a rather complex
reason but has a simple cause. Not going to bother to explain it
but the prototype chain is not recreated for an item and is the
same for every instance. Now this can help make some things more
efficient but hopefully this relization will help simplify some
code. The problem is such a change would be rather large which
would ubiquitously render a large portion of code invalid, I'll
look at changing this soon.

Fixes #66