hotfix [op]: Fixed equality operator with falsey values
With falsey primitives. `LHS.value` was added to prevent when
checking between non-primitives, to avoid false `true` values.
However this returned `false` when the LHS.value was `0` which
wasn't desired behavior. This was fixed by changing it to
`LHS.value !== undefined` which is stricter enforcment of the
desired behavior

Fixes #81

Signed-off-by: Vihan <contact@vihan.org>