feature [stdlib]: Added dynamic STDLIB  items
Within any STDLIB item. the `__COMPILE_SRC` and `__COMPILE` macros
will convert either the given file name, or string, respecively,
to a static, compiled cheddar CST. This can be passed to `api.src`
which will interpreter it additonally given the class name to
ignore, to avoid a cyclic dependency.

The cheddar file by convention shall be placed within a
`./natives.cheddar` directory. The very first line should be in the
format of:

    # itemName
    code

Where itemName will be the object's name in STDLIB. Additionally,
the `.src` function will look for a `main` item in the global scope
meaning a class, function, etc. can all be created

Signed-off-by: Vihan B <contact@vihan.org>