version [bump]: Set to release v1.0.0
Version v1.0.0 is the first major version of cheddar and brings an
entire language to the board. From dynamic code generation and
execution, to pre-compilation and classes, Cheddar v1 has been in
development for almost a 8 months and is finally ready for release.
Although a singificant backlog occured. A sound foundation for
scopes and classes have already made cheddar a powerful, extensible
language.

The changelist in this version is rather long but the highlights
include:

 - Dynamic Code Execution
 - Code pre-compilation
 - OO-Classes
 - Operator Overloading
 - User defined operators
 - Lambdas
 - Functional Operators
 - Functional Constructs
 - Symbols
 - Named Classes
 - Efficient recursion
 - Looping Constructs
 - A standard library
    - Powerful string manipulation
    - Powerful array managment
    - Math library, functions, and constants
    - An IO system
    - String formatting
    - Rational numbers
    - Wide-variety of encodings

Which is only a portion of the immense update Cheddar v1 is. Also,
a huge credit goes to all of the collaborators:

 - @ConorOBrien-Foxx: Pretty much all functional aspects of cheddar
 - @The-Quill: Unit testing
 - @LegionMammal978: Volunteering to make the Rational library
 - @somebody1234: Parser

This also does not include the large list of people who have gave
their valuable input, and advice on Cheddar's design (even through
strawpoll ;))

Signed-off-by: Vihan B <contact@vihan.org>